+++ packages/ionic/src/commands/config/index.ts
@@ -9,7 +9,7 @@ export class ConfigNamespace extends Namespace {
       name: 'config',
       description: 'Manage CLI and project config values',
       longDescription: `
-These commands are used to programmatically read and write CLI and project config values.
+These commands are used to programmatically read, write, and delete CLI and project config values.
 
 By default, these commands use your project's ${chalk.bold(PROJECT_FILE)} file.
 
@@ -22,6 +22,7 @@ To use these commands for the global CLI config file (${chalk.bold('~/.ionic/con
     return new CommandMap([
       ['get', async () => { const { ConfigGetCommand } = await import('./get'); return new ConfigGetCommand(this, this.env); }],
       ['set', async () => { const { ConfigSetCommand } = await import('./set'); return new ConfigSetCommand(this, this.env); }],
+      ['unset', async () => { const { ConfigUnsetCommand } = await import('./unset'); return new ConfigUnsetCommand(this, this.env); }],
     ]);
   }
 }